backend/pyproject.toml
[project]
dependencies = [
    "weave>=0.50.9",
    "openai>=1.12.0",
    "ollama>=0.1.7",
    "itsdangerous>=2.0.1",
    "peewee>=3.0.0",
    "fastapi>=0.111.0",
    "uvicorn>=0.22.0",
    "fastapi-sso>=0.10.0",
    "boto3>=1.34.67",
    "tiktoken>=0.8.0",
    "pandas>=2.3.0",
    "python-dotenv>=1.0.0",
]
name = "openui"
version = "0.5.0"
description = "A backend service for generating HTML components with LLM's"
readme = "README.md"
requires-python = ">=3.9"
authors = [{ name = "Chris Van Pelt" }]
license = { text = "Apache-2.0" }
classifiers = ["License :: OSI Approved :: Apache Software License"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
include = ["openui/**/*.py", "openui/*.yaml", "openui/*.ascii", "openui/assets/*", "openui/dist/*"]
exclude = [
  "openui/eval/components/*",
  "openui/eval/datasets/*",
  "openui/eval/wandb/*"
]

[project.urls]
Homepage = "https://github.com/wandb/openui"
Changelog = "https://github.com/wandb/openui/releases"
Issues = "https://github.com/wandb/openui/issues"
CI = "https://github.com/wandb/openui/actions"


[project.optional-dependencies]
test = ["pytest", "pre-commit"]
litellm = ["litellm[proxy]>=1.40.20", "google-generativeai"]
eval = [
    "beautifulsoup4>=4.0.0",
    "mistletoe>=1.0.0",
    "Pillow>=8.3.1",
    "playwright>=1.41.0",
]
tui = ["textual[syntax]>=0.49.0", "pyperclip>=1.8.2"]
